# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistem Pametne Kuće API",
        "description": "API dokumentacija za upravljanje uređajima, korisnicima i ulogama (Admin, Običan, Dete).",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Lokalni razvojni server"
        }
    ],
    "paths": {
...
        "/api/stanovi": {
            "get": {
                "tags": [
                    "Objekti (Stanovi)"
                ],
                "summary": "Prikaz svih objekata korisnika (gde je vlasnik ili stanar)",
                "operationId": "63dd2aea63397ccdbcd4c31b1199a497",
                "responses": {
                    "200": {
                        "description": "Uspešno vraćeni podaci o objektima"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Objekti (Stanovi)"
                ],
                "summary": "Dodaj novi stambeni objekat",
                "operationId": "97748c0c151604e8e1f33f7282019dc3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "adresa": {
                                        "type": "string",
                                        "example": "Bulevar Kralja Aleksandra 12"
                                    },
                                    "brojStana": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "sprat": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Objekat kreiran"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/stanovi/{id}": {
            "get": {
                "tags": [
                    "Objekti (Stanovi)"
                ],
                "summary": "Prikaz detalja objekta",
                "operationId": "a02c6da9952b76c6a9917511165f9309",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno vraćen objekat"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Objekti (Stanovi)"
                ],
                "summary": "Izmena podataka o objektu",
                "operationId": "a90a1a9268615a6efe22a0d9ea836fc7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno ažurirano"
                    },
                    "403": {
                        "description": "Niste vlasnik objekta"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Objekti (Stanovi)"
                ],
                "summary": "Obriši objekat",
                "operationId": "595ad0a68c70a79bd4a57a470f130181",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Objekat obrisan"
                    },
                    "404": {
                        "description": "Objekat nije pronađen"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/stanovi/{idStan}/ukloni-stanara": {
            "post": {
                "tags": [
                    "Objekti (Stanovi)"
                ],
                "summary": "Ukloni stanara iz objekta",
                "operationId": "3af2cde9c043592c8da222608bab8146",
                "parameters": [
                    {
                        "name": "idStan",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "korisnik_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stanar uklonjen"
                    },
                    "403": {
                        "description": "Niste vlasnik objekta"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/stanovi/{idStan}/dodaj-stanara": {
            "post": {
                "tags": [
                    "Objekti (Stanovi)"
                ],
                "summary": "Dodaj stanara u objekat preko korisničkog imena",
                "operationId": "24c8fef445d05da0f1e3bc9c9322ac00",
                "parameters": [
                    {
                        "name": "idStan",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "marko123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stanar uspešno dodat"
                    },
                    "403": {
                        "description": "Niste vlasnik objekta"
                    },
                    "404": {
                        "description": "Korisnik ne postoji"
                    },
                    "422": {
                        "description": "Korisnik je već stanar"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Ovde unesite access_token dobijen nakon logovanja.",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}